+++ src/providers/documentHighlightProvider.ts
@@ -8,6 +8,7 @@ import * as vscode from 'vscode';
 import * as parser from '../parsers/parser';
 import { Quest } from '../language/quest';
 import { TemplateReferenceProvider } from './referenceProvider';
+import { QuestResource } from '../language/common';
 
 export class TemplateDocumentHighlightProvider implements vscode.DocumentHighlightProvider {
 
@@ -22,13 +23,25 @@ export class TemplateDocumentHighlightProvider implements vscode.DocumentHighlig
                 // Symbol
                 const symbol = quest.qbn.getSymbol(word);
                 if (symbol) {
-                    const references = TemplateReferenceProvider.symbolReferences(quest, symbol, true);
-                    const highlights = references.map(x => new vscode.DocumentHighlight(x.range, vscode.DocumentHighlightKind.Read));
-                    return resolve(highlights);
+                    return resolve(TemplateDocumentHighlightProvider.makeHighlights(symbol,
+                        TemplateReferenceProvider.symbolReferences(quest, symbol, false)));
+                }
+
+                // Task
+                const task = quest.qbn.getTask(word);
+                if (task) {
+                    return resolve(TemplateDocumentHighlightProvider.makeHighlights(task,
+                        TemplateReferenceProvider.taskReferences(quest, task, false)));
                 }
             }
 
             return reject();
         });
+    }
+
+    private static makeHighlights(resource: QuestResource, references: vscode.Location[]) {
+        const highlights = references.map(x => new vscode.DocumentHighlight(x.range, vscode.DocumentHighlightKind.Read));
+        highlights.push(new vscode.DocumentHighlight(resource.range, vscode.DocumentHighlightKind.Write));
+        return highlights;
     }
 }